+++ lib/decompose.d.ts
@@ -1,10 +1,9 @@
-import type { types } from '@bemedev/types';
-import type { Decompose } from './types.types';
-type Decompose_F = <T extends types.TrueObject>(val: T) => Decompose<T>;
-type _Decompose_F = (val: any) => any;
+import { type Decompose, type DecomposeOptions } from './types.types';
+type Decompose_F = <T>(val: T, options?: DecomposeOptions) => Decompose<T>;
+type _Decompose_F = (val: any, options?: DecomposeOptions) => any;
 export type Decomposer = Decompose_F & {
-  strict: Decompose_F;
-  low: _Decompose_F;
+    strict: Decompose_F;
+    low: _Decompose_F;
 };
 export declare const decompose: Decomposer;
 export {};