+++ lib/recompose.d.ts
@@ -23,8 +23,8 @@ export declare function recomposeObjectUrl<T>(shape: string, value: T): Ru;
 type Recompose_F = <T extends Ru>(shape: T) => Recompose<T>;
 type _Recompose_F = (shape: any) => any;
 export type Recomposer = Recompose_F & {
-  strict: Recompose_F;
-  low: _Recompose_F;
+    strict: Recompose_F;
+    low: _Recompose_F;
 };
 export declare const recompose: Recomposer;
 export {};